refactor(rfid): separate rfid finished production warehouse management module
