+++ src/app/(features)/_layout.finished-production-inoutbound/_apis/rfid.api.ts
@@ -6,14 +6,14 @@ import { InoutboundPayload } from '../_schemas/epc-inoutbound.schema'
 import { type ExchangeEpcPayload } from '../_schemas/exchange-epc.schema'
 
 // * API Query Keys
-export const INOUTBOUND_DEPT_PROVIDE_TAG = 'SHAPING_DEPARTMENT'
-export const ORDER_DETAIL_PROVIDE_TAG = 'ORDER_DETAIL'
-export const EPC_LIST_PROVIDE_TAG = 'EPC_LIST'
+export const SHAPING_DEPT_PROVIDE_TAG = 'SHAPING_DEPARTMENT'
+export const FP_ORDER_DETAIL_PROVIDE_TAG = 'ORDER_DETAIL'
+export const FP_EPC_LIST_PROVIDE_TAG = 'FP_EPC_LIST'
 
 // * Fallback order value if it's null
 export const FALLBACK_ORDER_VALUE = 'Unknown'
 
-export type FetchEpcQueryKey = [typeof EPC_LIST_PROVIDE_TAG, number, string]
+export type FetchEpcQueryKey = [typeof FP_EPC_LIST_PROVIDE_TAG, number, string]
 
 export const useGetEpcQuery = () => {
 	const { currentPage, selectedOrder, connection, scanningStatus } = usePageContext(
@@ -24,8 +24,8 @@ export const useGetEpcQuery = () => {
 	)
 
 	return useQuery({
-		queryKey: [EPC_LIST_PROVIDE_TAG, connection],
-		queryFn: async () => RFIDService.fetchEpcManually(connection, currentPage, selectedOrder),
+		queryKey: [FP_EPC_LIST_PROVIDE_TAG, connection],
+		queryFn: async () => RFIDService.fetchFPData(connection, currentPage, selectedOrder),
 		enabled: !!connection && scanningStatus === 'disconnected',
 		select: (response) => response.metadata
 	})
@@ -35,7 +35,7 @@ export const useGetOrderDetail = () => {
 	const { connection, scanningStatus } = usePageContext('connection', 'scanningStatus')
 
 	return useQuery({
-		queryKey: [ORDER_DETAIL_PROVIDE_TAG, connection],
+		queryKey: [FP_ORDER_DETAIL_PROVIDE_TAG, connection],
 		queryFn: async () => await RFIDService.getOrderDetail(connection),
 		enabled: !!connection && scanningStatus === 'disconnected',
 		select: (response) => response.metadata
@@ -55,7 +55,7 @@ export const useSearchOrderQuery = (orderTarget: string, searchTerm: string) => 
 
 export const useGetShapingProductLineQuery = () => {
 	return useQuery({
-		queryKey: [INOUTBOUND_DEPT_PROVIDE_TAG],
+		queryKey: [SHAPING_DEPT_PROVIDE_TAG],
 		queryFn: DepartmentService.getShapingDepartments,
 		select: (response) => response.metadata
 	})
@@ -71,7 +71,7 @@ export const useDeleteOrderMutation = () => {
 
 	return useMutation({
 		mutationFn: async (orderCode: string) => {
-			return await RFIDService.deleteUnexpectedOrder(connection, orderCode)
+			return await RFIDService.deleteUnexpectedFPOrder(connection, orderCode)
 		},
 		onSuccess: () => {
 			setCurrentPage(null)
@@ -123,13 +123,13 @@ const useInvalidateQueries = () => {
 
 	return () => {
 		queryClient.invalidateQueries({
-			queryKey: [ORDER_DETAIL_PROVIDE_TAG, connection],
+			queryKey: [FP_ORDER_DETAIL_PROVIDE_TAG, connection],
 			exact: false,
 			type: 'all',
 			refetchType: 'all'
 		})
 		queryClient.invalidateQueries({
-			queryKey: [EPC_LIST_PROVIDE_TAG, connection],
+			queryKey: [FP_EPC_LIST_PROVIDE_TAG, connection],
 			exact: false,
 			type: 'all',
 			refetchType: 'all'
