+++ src/app/(features)/_layout.finished-production-inoutbound/_components/_manufacturing-order-detail/-index.tsx
