+++ src/app/(features)/_layout.finished-production-inoutbound/_components/_scanner-settings/-setting-panel.tsx
