+++ src/app/(features)/_layout.finished-production-inoutbound/_constants/rfid.const.ts
@@ -1,8 +1,3 @@
-/**
- * @event
- */
-export const INCOMING_DATA_CHANGE = 'INCOMING_DATA_CHANGE' as const
-
 export enum RFIDSettings {
 	SSE_POLLING_DURATION = 'rfidSettings.ssePollingDuration',
 	PRESERVE_LOG = 'rfidSettings.preserveLog',
