+++ src/common/errors/index.ts
@@ -1,0 +1,2 @@
+export class RetriableError extends Error {}
+export class FatalError extends Error {}
